Remove console.log e move ShoppingCart de diretorio
