--- src/components/ShoppingCart.js
@@ -1,11 +1,0 @@
-import React, { Component } from 'react';
-
-class ShoppingCart extends Component {
-  render() {
-    return (
-      <p data-testid="shopping-cart-empty-message">Seu carrinho está vazio</p>
-    );
-  }
-}
-
-export default ShoppingCart;
